pyproject.toml
[project]
name = "mailcast"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
authors = [
	{ name = "Felix C.", email = "11308118@gs.ncku.edu.tw" },
]
dependencies = [
	"html-to-markdown[lxml]>=1.9.0",
	"icalendar>=6.3.1",
	"jinja2>=3.1.6",
	"jsonpath-rfc9535>=0.1.6",
	"loguru>=0.7.3",
	"pydantic-extra-types>=2.10.5",
	"pydantic[email]>=2.11.7",
	"python-dotenv>=1.1.0",
	"pyyaml>=6.0.2",
	"rich>=14.1.0",
	"typer>=0.17.3",
]

[project.urls]
#

[project.optional-dependencies]
#

[project.scripts]
mailcast = "mailcast.cli.__init__:main"

[dependency-groups]
dev = [
	"ruff>=0.12.10",
	"types-icalendar>=6.3.1.20250809",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
exclude = ["dist"]
line-length = 120
respect-gitignore = true
target-version = "py313"

[tool.ruff.format]
indent-style = "space"
line-ending = "lf"
quote-style = "double"
skip-magic-trailing-comma = false

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]

[tool.ruff.lint.isort]
combine-as-imports = true
force-single-line = false
force-sort-within-sections = false
force-wrap-aliases = true
split-on-trailing-comma = true
